+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Macros
 
 - Error diagnostics is improved
 - `MemberwiseInitializable` will no longer throw errors when empty `Dictionary` is used for `externalParameterNames` and `parameterDefaultValues` parameters
+- `"*": .omit` wildcard can now be used in `parameterDefaultValues` in `MemberwiseInitializable` to remove all default values from parameters
 
 ### [7.4.0(108)](https://github.com/VakhoKontridze/VCore/releases/tag/7.4.0) — *2025 01 28*